requirements.txt
docxtpl==0.16.7
Flask==3.0.0
python_docx==1.1.0
pywin32==305
